End_sem/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pandas
numpy
openpyxl
scikit-learn
joblib
plotly
streamlit
langchain
langchain-core
langchain-community
langchain-openai
langchain-huggingface
langchain-text-splitters
langgraph
sentence-transformers
faiss-cpu
python-multipart
